Create Smart Spell alternator
Small addition for your vimrc for all you bilingual folks who need to change the language of spelling vim is using, without having to use them both at the same time. With this binding, Ctrl+i should alternate between the two languages (A and B in the code), provided you have the files for both of them installed. This would be the equivalent to WIN+SPACE but exclusively for vim